requirements.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
Django==5.2.8
djangorestframework==3.16.1
idna==3.11
nats-py==2.12.0
nkeys==0.2.1
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
PyNaCl==1.6.1
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.3
tmdbsimple==2.9.1
urllib3==2.5.0
